requirements.txt
# Satellite imagery + geospatial
pystac-client>=0.7.5
planetary-computer>=1.0.0
odc-stac>=0.3.8
rioxarray>=0.15.0
xarray>=2023.12.0
rasterio>=1.3.9
geopandas>=0.14.0
shapely>=2.0.0

# Data + math
numpy>=1.26.0
pandas>=2.1.0
pillow>=10.0.0

# Visualization + UI
streamlit>=1.42.0
streamlit-folium>=0.22.0
folium>=0.15.0
matplotlib>=3.8.0
plotly>=5.18.0

# Misc
requests>=2.31.0
python-dateutil>=2.8.2
